C5 configuration model: phase 05
