--- configuration/annotations/pom.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <groupId>org.wso2.carbon</groupId>
-        <artifactId>org.wso2.carbon.configuration</artifactId>
-        <version>5.2.0-SNAPSHOT</version>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-
-    <artifactId>org.wso2.carbon.configuration.annotations</artifactId>
-    <packaging>bundle</packaging>
-    <name>WSO2 Carbon Configuration - Annotations</name>
-    <description>
-        Define annotations used for carbon configurations
-    </description>
-    <url>http://wso2.com</url>
-
-    <properties>
-        <export.package>
-            org.wso2.carbon.configuration.annotations.*; version="${project.version}"
-        </export.package>
-    </properties>
-</project>
