--- configuration/pom.xml
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-<!--    <parent>
-        <groupId>org.wso2.carbon</groupId>
-        <artifactId>carbon-kernel-parent</artifactId>
-        <version>5.2.0-SNAPSHOT</version>
-    </parent>-->
-    <parent>
-        <groupId>org.wso2.carbon</groupId>
-        <artifactId>carbon-kernel-parent</artifactId>
-        <version>5.2.0-SNAPSHOT</version>
-        <relativePath>../parent/pom.xml</relativePath>
-    </parent>
-
-    <modelVersion>4.0.0</modelVersion>
-    <artifactId>org.wso2.carbon.configuration</artifactId>
-    <name>WSO2 Carbon Kernel - Configuration</name>
-    <modules>
-        <module>annotations</module>
-        <module>annotation-processors</module>
-        <module>annotation-plugins</module>
-    </modules>
-    <packaging>pom</packaging>
-    <description>
-        This is the module for all the Carbon configurations
-    </description>
-    <url>http://wso2.com</url>
-
-
-</project>
