+++ core/pom.xml
@@ -92,17 +92,6 @@
             <artifactId>powermock-module-testng</artifactId>
             <scope>test</scope>
         </dependency>
-        <dependency>
-            <groupId>org.wso2.carbon</groupId>
-            <artifactId>org.wso2.carbon.configuration.annotations</artifactId>
-            <version>5.2.0-SNAPSHOT</version>
-        </dependency>
-        <dependency>
-            <groupId>org.wso2.carbon</groupId>
-            <artifactId>org.wso2.carbon.configuration.annotations.processors</artifactId>
-            <version>5.2.0-SNAPSHOT</version>
-            <scope>compile</scope>
-        </dependency>
     </dependencies>
 
     <build>
@@ -167,19 +156,6 @@
                     </execution>
                 </executions>
             </plugin>
-            <plugin>
-                <groupId>org.wso2.carbon</groupId>
-                <artifactId>org.wso2.carbon.configuration.annotations.plugins</artifactId>
-                <version>5.2.0-SNAPSHOT</version>
-                <executions>
-                    <execution>
-                        <goals>
-                            <goal>create-doc</goal>
-                        </goals>
-                        <phase>compile</phase>
-                    </execution>
-                </executions>
-            </plugin>
         </plugins>
     </build>
 
@@ -204,8 +180,7 @@
             org.w3c.*;version="${org.w3c.import.version.range}",
             org.json.*;version="${org.json.import.version}",
             org.xml.sax.*;version="${org.xml.sax.import.version.range}",
-            javax.crypto.*;version="${javax.crypto.version.range}",
-            org.wso2.carbon.configuration.annotations.*; version="${project.version}"
+            javax.crypto.*;version="${javax.crypto.version.range}"
         </import.package>
         <carbon.component>
             startup.listener;componentName="carbon-transport-mgt";requiredService="org.wso2.carbon.kernel.transports.CarbonTransport",
