+++ core/src/main/java/org/wso2/carbon/kernel/config/model/DeploymentConfig.java
@@ -16,8 +16,8 @@
 package org.wso2.carbon.kernel.config.model;
 
 
-import org.wso2.carbon.configuration.annotations.Configuration;
-import org.wso2.carbon.configuration.annotations.Element;
+import org.wso2.carbon.kernel.annotations.Configuration;
+import org.wso2.carbon.kernel.annotations.Element;
 
 /**
  * Deployment Config bean.
