+++ core/src/main/java/org/wso2/carbon/kernel/config/model/StartupResolverConfig.java
@@ -18,7 +18,7 @@
 package org.wso2.carbon.kernel.config.model;
 
 
-import org.wso2.carbon.configuration.annotations.Configuration;
+import org.wso2.carbon.kernel.annotations.Configuration;
 
 /**
  * Config bean for startupOrderResolver in carbon.yaml file.
