+++ features/org.wso2.carbon.kernel.feature/pom.xml
@@ -36,11 +36,6 @@
         </dependency>
         <dependency>
             <groupId>org.wso2.carbon</groupId>
-            <artifactId>org.wso2.carbon.configuration.annotations</artifactId>
-            <version>5.2.0-SNAPSHOT</version>
-        </dependency>
-        <dependency>
-            <groupId>org.wso2.carbon</groupId>
             <artifactId>org.wso2.carbon.base</artifactId>
         </dependency>
         <dependency>
@@ -170,10 +165,6 @@
                             <bundles>
                                 <bundle>
                                     <symbolicName>org.wso2.carbon.core</symbolicName>
-                                    <version>${carbon.kernel.version}</version>
-                                </bundle>
-                                <bundle>
-                                    <symbolicName>org.wso2.carbon.configuration.annotations</symbolicName>
                                     <version>${carbon.kernel.version}</version>
                                 </bundle>
                                 <bundle>
