+++ features/org.wso2.carbon.kernel.feature/resources/conf/deployment.yaml
@@ -1,61 +1,64 @@
 ################################################################################
-#   Copyright (c) 2015, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+#   Copyright (c) 2016, WSO2 Inc. (http://www.wso2.org) All Rights Reserved
 #
-#   Licensed under the Apache License, Version 2.0 (the "License");
+#   Licensed under the Apache License, Version 2.0 (the \"License\");
 #   you may not use this file except in compliance with the License.
 #   You may obtain a copy of the License at
 #
 #   http://www.apache.org/licenses/LICENSE-2.0
 #
 #   Unless required by applicable law or agreed to in writing, software
-#   distributed under the License is distributed on an "AS IS" BASIS,
+#   distributed under the License is distributed on an \"AS IS\" BASIS,
 #   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 ################################################################################
 
-id: ${server.key}           #Value to uniquely identify a server
-name: ${server.name}        #Server Name
-version: ${server.version}  #Server Version
-tenant: default      #Tenant Name
-
-# Ports used by this server
-ports:
-# Ports offset. This entry will set the value of the ports defined below to
-# the define value + Offset.
-# e.g. Offset=2 and HTTPS port=9443 will set the effective HTTPS port to 9445
- offset: ${server.offset}
-# Deployment Engine related configurations
-deployment:
-# Currently there can be two modes
-#  1. Scheduled Mode - (enable below property to "scheduled") - where the task runs periodically and trigger deployment
-#  2. Triggered Mode - (enable below property to "triggered") - the deployment has to be triggered externally,
-#   eg : in a worker node we don't need the task to run, but rather when we receive a
-#        cluster msg,  the deployment has to be triggered manually
- mode: scheduled
-
-# Location of the artifact repository
- repositoryLocation: ${carbon.home}/deployment/
-
-# Deployment update interval in seconds. This is the interval between repository listener
-# executions.
- updateInterval: 15
-
-# StartupOrderResolver related configurations
-startupResolver:
-# Configuration for the timer task which checks for satisfiable RequiredCapabilityListeners periodically
- capabilityListenerTimer:
-  delay: 20   #delay in milliseconds before task is to be executed
-  period: 20  #time in milliseconds between successive task executions
-
-# Configuration for the timer task which checks for pending Capabilities
- pendingCapabilityTimer:
-  delay: 5000    #delay in milliseconds before task is to be executed
-  period: 5000   #time in milliseconds between successive task executions
-
-# JMX Configuration
-jmx:
- enabled: false         #To enable JMX Monitoring, change this value to true
- hostName: "127.0.0.1"  #Server HostName
- rmiServerPort: 11111   #The port RMI server should be exposed
- rmiRegistryPort: 9999  #The port RMI registry is exposed
+  # Carbon Configuration Parameters
+wso2.carbon:
+    # value to uniquely identify a server
+    # THIS IS A MANDATORY FIELD
+  id: carbon-kernel
+    # server name
+  name: WSO2 Carbon Kernel
+    # server version
+  version: 5.2.0-SNAPSHOT
+    # ports used by this server
+  ports:
+      # port offset
+    offset: 0
+    # deployment engine related configurations
+  deployment:
+      # Currently there can be two modes
+      #  1. Scheduled Mode - (enable below property to "scheduled") - where the task runs periodically and trigger deployment
+      #  2. Triggered Mode - (enable below property to "triggered") - the deployment has to be triggered externally,
+      #   eg : in a worker node we don't need the task to run, but rather when we receive a cluster msg,
+      # the deployment has to be triggered manually.
+    mode: scheduled
+      # Location of the artifact repository
+    repositoryLocation: ${carbon.home}/deployment/
+      # Deployment update interval in seconds. This is the interval between repository listener executions.
+    updateInterval: 15
+    # StartupOrderResolver related configurations
+  startupResolver:
+    capabilityListenerTimer:
+        # delay in milliseconds before task is to be executed
+      delay: 20
+        # time in milliseconds between successive task executions
+      period: 20
+    pendingCapabilityTimer:
+        # delay in milliseconds before task is to be executed
+      delay: 60000
+        # time in milliseconds between successive task executions
+      period: 30000
+    # JMX Configuration
+  jmx:
+      # To enable JMX Monitoring, change this value to true
+    enabled: false
+      # Server HostName
+    hostName: 127.0.0.1
+      # The port RMI server should be exposed
+    rmiServerPort: 11111
+      # The port RMI registry is exposed
+    rmiRegistryPort: 9999
+    # Testing String array
