+++ plugins/configuration-plugin/src/main/resources/LICENSE.txt
